fix imports referencing moved metrics.py file
