+++ examples/adaptive_span/adaptive_span_loss.py
@@ -7,7 +7,8 @@ import math
 from dataclasses import dataclass
 
 import torch.nn.functional as F
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import register_criterion
 from fairseq.criterions.cross_entropy import CrossEntropyCriterion
 from fairseq.dataclass import FairseqDataclass
