+++ examples/discriminative_reranking_nmt/criterions/discriminative_reranking_criterion.py
@@ -9,7 +9,8 @@ from dataclasses import dataclass, field
 import torch
 import torch.nn.functional as F
 
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import FairseqCriterion, register_criterion
 from fairseq.dataclass import ChoiceEnum, FairseqDataclass
