+++ fairseq/criterions/fairseq_criterion.py
@@ -6,7 +6,8 @@
 import inspect
 from typing import Any, Dict, List
 
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.dataclass import FairseqDataclass
 from fairseq.dataclass.utils import gen_parser_from_dataclass
 from torch.nn.modules.loss import _Loss
