+++ fairseq/criterions/label_smoothed_cross_entropy.py
@@ -9,7 +9,8 @@ from dataclasses import dataclass, field
 import torch
 from omegaconf import II
 
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import FairseqCriterion, register_criterion
 from fairseq.dataclass import FairseqDataclass
