+++ fairseq/criterions/speech_to_speech_criterion.py
@@ -9,7 +9,8 @@ from collections import OrderedDict
 
 import torch
 
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import register_criterion
 from fairseq.criterions.ctc import CtcCriterion
 from fairseq.criterions.label_smoothed_cross_entropy_with_rdrop import (
