+++ fairseq/model_parallel/criterions/vocab_parallel_cross_entropy.py
@@ -5,7 +5,8 @@
 
 import math
 
-from fairseq import metrics, utils
+from fairseq import utils
+from fairseq.logging import metrics
 from fairseq.criterions import FairseqCriterion, register_criterion
