+++ fairseq/tasks/online_backtranslation.py
@@ -20,7 +20,8 @@ import torch.nn as nn
 import torch.nn.functional as F
 
 import fairseq
-from fairseq import metrics, options, utils
+from fairseq import options, utils
+from fairseq.logging import metrics
 from fairseq.data import (
     FairseqDataset,
     LanguagePairDataset,
